Moved output stencil code
